Problem solved, clean the mess
the problem was : I was putting my output in a dedicated "output" folder. Which  by default is empty. And it seems that git/github doesn't "see" empty folders? it exeists only if there is something in it.

So locally I had this empty folder ready to receive files, but when copying the repo into the docker I guess this wasn't taken into account

Problem solved by forcing the creatioin of the saiid folder